Update dependencies to latest; raise MSRV to 1.88
Bump the lagging majors (symphonia 0.5->0.6, quick-xml 0.39->0.40,
object 0.36->0.39, rusqlite 0.32->0.40, sha1/sha2 0.10->0.11,
lz4_flex 0.11->0.13, windows-sys 0.59->0.61) plus a cargo update sweep.
image/tiff/time/plist now require Rust 1.88, so the floor follows them.

API migrations the bumps forced:
- symphonia 0.6 reworked codecs (AudioCodecId + well_known module),
  metadata (StandardTag enum, RawValue, rev.media container) and the
  probe (probe() -> FormatReader, TrackType-keyed default_track).
  Codec name strings and extracted fields are unchanged.
- quick-xml 0.40 made xml_content() take an XmlVersion; switch the 13
  call sites to xml10_content() (XML 1.0, prior behaviour).
- object 0.39 widened DT_* to i64; narrow DT_NEEDED for the i32 compare.

Regenerate test-audio/sample-tagged.mp3 with 8 frames: symphonia 0.6's
MPEG probe confirms the format by reading a second sync word, which the
old single-frame fixture lacked.

+++ crates/peek-types/src/types/objfile/links.rs
@@ -44,7 +44,9 @@ fn elf_needed<Elf: FileHeader<Endian = Endianness>>(data: &[u8]) -> Vec<String> 
         .unwrap_or_default();
     entries
         .iter()
-        .filter(|d| d.tag32(endian) == Some(elf::DT_NEEDED))
+        // object 0.39 widened the `DT_*` constants to `i64`; `tag32`
+        // still yields `Option<i32>`, so narrow the constant to compare.
+        .filter(|d| d.tag32(endian) == Some(elf::DT_NEEDED as i32))
         .filter_map(|d| d.string(endian, strings).ok())
         .map(|name| String::from_utf8_lossy(name).into_owned())
         .collect()
